daemon/pyproject.toml
[project]
name = "amphigory-daemon"
version = "0.1.0"
description = "macOS host daemon for Amphigory media processing"
requires-python = ">=3.11"
dependencies = [
    "rumps>=0.4.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "PyYAML>=6.0",
    "pyobjc-framework-Cocoa>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
amphigory-daemon = "amphigory_daemon.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
